requirements.txt
Django==2.2.2
django-webpack-loader==0.6.0
djangorestframework==3.9.4
